Elemental affinity
Added elemental affinity to magic powers. When someone casts a
magic power, the current element affinity is to the element of
that magic. If the player is hit by the same element, it absorbs
part of the damage or some effect of that power. If the element is
the opposite one, instead the player takes more damage and more of
some other effects of that magic.