Quest items
The green crystal will now be one of the parts of a new secret
item called Spirit Crystal. It will have the abilities that were
given by the Quest Log. The Quest Log will now only give info. The
red crystal will be the one that regens the Spirit Tree.